feat(mobile): add social entry list component (#2733)
* refactor(mobile): extract entry list article component

* feat(mobile): add social entry list component

* fix: lint

* chore: typo

* refactor(mobile): replace FlashList with TimelineSelectorList

* chore: auto-fix linting and formatting issues

* refactor(mobile): extract ItemSeparator component

* chore: add diff

* fix: merge

---------
