+++ apps/mobile/src/modules/entry-list/ItemSeparator.tsx
@@ -1,0 +1,9 @@
+import { StyleSheet, View } from "react-native"
+
+const el = (
+  <View className="bg-opaque-separator mx-4" style={{ height: StyleSheet.hairlineWidth }} />
+)
+
+export const ItemSeparator = () => {
+  return el
+}
